packaging: Install the initrd and vmlinuz from another rpm
Makefile contains additional targets to help with packaging. Now the
initrd and vmlinuz files are extracted from separate SRPM
  leapp-repository-initrd
as a tarball. The SRPM should be part of the same repository in COPR.
The tarball is used as additional SOURCE in the SPEC file and files
are copied to proper path during the build phase.