Change the repos hierarchy: initial commit
The original directory hierarchy was not suitable for repositories
supposed to be used for various systems. Current hierarchy should
be sufficient for future to support repositories for various systems
under the one upstream.

The new proposed schema of hierarchy:
    type_of_repos/<src_system>to<dst_system>
    common/

So for the system upgrade from RHEL 7 to RHEL 8:
    system_upgrade/el7toel8

E.g. the hierarchy would look like:
    .
    ├── common
    ├── containarization
    |   ├── el7toel8
    |   ├── centos7tocentos8
    |   ├── f29tof30
    |   └── ...
    ├── system_upgrade
    |   ├── el7toel8
    |   ├── centos7tocentos8
    |   ├── f29tof30
    |   └── ...
    └── ...

In addition the offline-upgrade and the upgrade repositories have
been merged together, as there is only one type of upgrade supported
now. The merged repository is called *system_upgrade_el7toel8* as
proposed best practice in future will be to call repositories
regarding the used path under the repos directory.

TODO:
    - fix the SPEC file