Reduce the IPUWorkflow workflow
We found that several phases are not usable from our perspective
of in-place upgrades. Such phases are commented out to make clear
they should not be used and keeping info what was our original idea
in case we think about them again in future.

Remove tags of avoided phases of IPUWorkflow.