Add metapackage for dependencies
Currently we are not able to use python dependencies correctly to
ensure the RPM will not be removed during the upgrade of the system.
As a workround, put all such system dependencies into the
leapp-repository-deps metapackage which is supposed to be replaced
later (another commit(s)) during the upgrade transaction by the
another metapackage containing dependencies satisfying dependencies
of target (RHEL 8) system.

To reflect way of the leapp framework how this should be resolved,
the leapp-repository rpm should create dependency on specific
capability:
  leapp-repository-dependencies

which has to be provided by such meta package. Any time the
dependencies are changed, the capability number has to be incremented
by one. Such change has to be reflected by both (for RHEL7 and RHEL8)
metapackages.