Add ntp migration actors
Add CheckNtp actor to check if the ntpd, ntpdate and ntp-wait services
are enabled, active and their configuration files exist. Save the
configuration files in a string containing a base64-encoded compressed
tarball to avoid having to deal with .rpmsave files when the packages
are removed in the rpm upgrade.

Add MigrateNtp actor to enable the chronyd and chrony-wait services,
restore the ntp/ntpdate configuration files, and run the included
ntp2chrony migration script. Produce a warning with a suggestion to
check the generated /etc/chrony.conf if some lines in ntp.conf were
ignored in the migration.