Support upgrade to authselect
AuthselectScanner reads current configuration and tries to map it to an authselect
call. See actor's doc block for details.

AuthselectCheck will ask the administrator if it is ok to convert the system to
authselect. The question is necessary in order to avoid undesirable changes as
it is not possible to predict any possible configuration. See actor's doc block
for details.

AuthselectApply will consume messages from the previos actor and if the change
was confirmed it will call authselect with suggested parameters on rhel8 system.
See actor's doc block for details.