SKIP RHSM (devel)
For the testing purposes we need to skip use of subscription-manager
and go on just with custom repositories. This is now possible with
the patch when proper custom repositories are set. The solution will
be reworked in future.

To skip the subscription-manager, set
  LEAPP_DEVEL_SKIP_RHSM=1

To be able to upgrade without RHSM, you need to deliver RHEL 8 repos
by creating actor that will produce `CustomTargetRepository` msg.
The repo file of such repositories has to exist.