Remove old PAM modules from configuration
pam_pkcs11 and pam_krb5 are no longer available in RHEL-8. Leaving these
modules in PAM configuration may lock out the system so they must be
unconditionaly removed.

RemoveOldPAMModulesScanner reads current PAM configuration (only standard files
/etc/pam.d/system-auth, password-auth, fingerprint-auth, smartcard-auth) and
produces a model with a list of modules that are present but no longer supported.

RemoveOldPAMModulesCheck will ask the administrator if the modules can be automatically
removed. If not, then the upgrade is aborted as keeping the modules in the configuration
would lock the system after upgrade.

RemoveOldPAMModulesApply will comment the modules in the configuration.