Handle the yum configuration files (Fixes #19)
RPM cannot handle replace of directories by symlinks by default
without the %pretrans scriptlet. As yum package is packaged wrong,
we have to workround that by actors so we will be able to keep
installed yum package.

Current solution handle the yum configuration files to replace
relevant directories by symlinks as expected on target system. This
change is done on two places:
- inside the container when RPM transactions is calculated and rpms
  are downloaded (the preupgradeupgradetransaction actor)
  - system is unaffected by this change
- before the upgrade itself (PreparationPhase) in the prepareyumconfig
  actor.

The important code handling that migration is stored in the tools
directory of the el7toel8 repository. See the handleyumconfig tool.