Start leapp with Python3 after RPMUpgradePhase
We found several issues using Python2 after the RPM transaction, e.g.
missing python libraries which are no longer provided for Python2 or
mixture of loaded libraries from RHEL 7 and RHEL 8 when run leapp
in one neverending process.

To resolve this, we will stop the IPUWorkflow after the RPM
transaction (RPMUpgradePhase) and restart it with Python3.
Additionally, initramdisk is modified as well to start the new
leapp script with python3 as expected.

To stop the processing of the workflow, use the is_checkpoint flag
which has been provided recently. Current solution is hacky as we
are creating directory with files inside under the /root/ - but
this is necessary hack now. We will try to come with better solution.

Added RPM dependency for the python3 package and bumped capability.