Fix the networkmanagerupdateconnections actor
The helper script to update client-ids needs a python D-Bus library
which is only available for Python 3 in the new system. Therefore,
explicitly run the script through the python3 interpreter.

Also, the script needs to be called when NetworkManager is running and
so move the actor to the FirstBoot phase.