OpenLDAP - migrate NSS to OpenSSL format
OpenLDAP in el7 used to use NSS as a crypto backend. However, this
changed in el7.5 (with a backwards compaibility support) to OpenSSL
which uses different format for storing certificates. With el8 there is
no more compatibility support for NSS database, therefore there is need
to convert old-style configuration to the new one. In el8 this affects
ldap.conf configuration file but also sssd.conf which uses libldap of
OpenLDAP project with its own configuration needing correction, too.