dracut: Ensure the $NEWROOT is set and use default otherwise
We hit *somtimes* on some machines problem that we are not even able
to mount the rootfs. In such cases, there is just err message like
       mount: bad usage
       Try 'mount --help' for more information.

It's weird, some modules checks the  $NEWROOT variable, some not.
So just from this point to ensure whether this would be the source
of the issue, putting here the initialisation when NEWROOT is not
defined.