Add actor for reading vsftpd configuration files
The main purpose of having the relatively complex parser is so that we
can be sure that we're really modifying a vsftpd config file and not
some other file ending with '.conf' that the user happened to have put
inside /etc/vsftpd.

In the VsftpdConfig model, the attributes representing the state of
configuration options are nullable, so that they can represent the real
state of the option in the file: if an option is set to "YES" in the
configuration file, the corresponding attribute is set to True; if
the option is set to NO, the attribute is set to False; if the option
is not present in the config file at all, the attribute is set to None.
Designed this way, the VsftpdConfigRead actor can be generic - it does
not need to know the default values of the configuration options, and
so the actor could be used to read the configuration of any version of
vsftpd. Said another way, the actor only reads the configuration files,
it does not interpret them.

v2: Simplified MockFile.read_file() in test_vsftpdutils.py
v3: Renamed private functions to begin with an underscore.
v4: Documented the public API in the public library.
v5: Fixed traceback when /etc/vsftpd/ is inaccessible, added a test for it.
v6: Fixed tests to raise OSError instead of IOError in mocked listdir.
    Fixed creating IOError and OSError in tests.
    Catch OSError instead of IOError when using listdir.
v7: Fixed the actor not to attempt to read vsftpd configs at all if
    vsftpd is not installed.
    Document the VsftpdConfig and VsftpdFacts models.