Add an actor for modifying the vsftpd config file
This actor changes the vsftpd config files so that they still work in
RHEL-8, and so that the effective configuration is as close to the
original as possible.

Namely, this actor explicitly sets the strict_ssl_read_eof option in
config files, if it was not set, because the default value has changed
between RHEL-7 and RHEL-8.

The actor also restores the value of the anonymous_enable option, if it
changed during the upgrade - this can happen if the
/etc/vsftpd/vsftpd.conf config file was unmodified on the source system,
and therefore the file got replaced with the new version during the RPM
upgrade.

The actor also sets the tcp_wrappers option to "NO", as tcp_wrappers is
not supported in RHEL-8.

v2: Renamed private functions to begin with an underscore.
v3: Fix creating IOError in tests.