rpmscanner: move get_package_repository_data to library
Because of `import yum` upgrades started to be broken as when we
start leapp under Python3 (after the rpmupgrade phase) leapp crashes
during the scanning of the repository because the yum module for
Python3 doesn't exist.

Moving the functionality into the actor's library as it is not
necessary to parse the code when it is not used (the function is
used just during the FactsCollectionPhase phase.