Makefile: fix building of packages
Previous solution worked just for copr-cli v1.78 and lower. Adding
the helper script that process data obtained from the COPR server
and provide expected content for scripts in Makefile, so building
will be possible with various versions of copr-cli.

Additionally, the helper script provides simple library that helps
with management of COPR builds in repository, which can be useful
in case someone would like to manage builds in COPR repository.