Don't join command arguments
The way the arguments are really passed to `run()` and underlying OS is
directly, without space or quotes having any specific meaning.

So joining the arguments with space actually makes things confusing
because now a valid list of args becomes an invalid string (see `--title
RHEL Upgrade...`).

It's safer and more relevant to compare the lists directly.