Introduce mounting library
This library provides functionality for:
- Bind mounts
- Loop mounts
- Overlayfs mounts
- Typed mounts e.g. sysfs or procfs etc

In a way that they automatically revert back when leaving the context.
This is done by context manager style operations.

On top of that, the library implements an isolation view layer to run
operations as if it were the current system.
Processes are run using systemd-nspawn or chroot, filesystem operations
wrap the execution to abstract away the need for joining paths when
accessing those isolated locations.

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>