Actor that generates the initial target userspace
The TargetUserSpaceCreator actor creates a directory where it installs
dnf with all its depedencies from the target system.

To be able to retrieve the packages from the target system it switches
the product certificate in an overlay over the real system, unsets the
release and refreshes the subscription. It will copy the certificates
and repository files to the userspace to make them available for later
usage. Then it will deploy dnf and all its dependencies to
/var/lib/leapp/el8userspace.

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>