Remove _logger helper functions
Previously the _logger helper function has been used in several private
and common libraries for convenience. However There was or is only
little value in it, and looks like a lot of code duplication. This patch
removes those helper functions and replaces it with the appropriate
api.current_logger function.