Set RHSM target release
Previously we did not set any explicit target version on RHSM which
could cause us to receive newer content than what actually is requested
for the upgrade.
We also did not unset the release after the upgrade, which could have
caused dnf update calls to fail, due to old version release value being
set. However if we unset the release and would not set the release the user
might end up updating the system to a newer version than the upgrade was
supposed to be upgrading to.

This PR solves this by setting first the release when we are setting up
all repositories and the target userspace and by setting the release in
the first boot phase after the upgrade has been completed.
Additionally it will let the user know about the changes applied, by
sending a report.

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>