dnfplugin: Make repos also obey gpgcheck = False
In case of upgrades, we usually set the gpgcheck to false. However the
repositories have their own values which need to be explicitly
overridden by the plugin too, if we don't want the gpgcheck to be
performed.
This patch ensures, that the gpgcheck value is set to false explicitly
if we are instructed to set the gpgcheck=False through the plugin
config.

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>