Blacklist the CRB repository when Optional repo is not enabled
Previously the CRB repository was blacklisted always. Now we will
blacklist this one only when the Optional repository is not enabled
as the CRB is sort of counterpart of the Optional repo in RHEL 8
system.

Additionally, the filtering of blacklisted repositories in
the peseventsscanner actor is applied after repositories are mapped
from aliases to real repoids.