Remove all richdeps RPMs from to_remove
Previously we did not support upgrades of RPMs where the RHEL8
counterpart was using RichDependencies.
Since we merged the RHEL8 userspace this feature should now be available
to us and therefore we are starting with a clean slate here.
Left are only packages that are in the list due to other reasons than
richdependencies.

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>