Enable CRB repository when used during the upgrade
We are using the CRB repository during the upgrade when the Optional
repository was enabled on the source system. As this one
is counterpart to the Optional repo, enable it as well on the target
system (during the FirstBoot).

TODO: the actor is working with this one repo now, but would be nice
      to make it more robust to be used for any repository task
      in future. See the TODO msg in code.