linter: enable invalid-name check
This removes invalid-name check from disabled
by defaut checks. Currently regexes are super
relaxed and any name is allowed, to fix this
consult [1].
Not making them strict by default as it will
require plenty of time and effort to fix existing
codebase, though in the long perspective that's
a nive thing to do.

[1] http://pylint-messages.wikidot.com/messages:c0103