Enable crb <=> the Optional repo is enabled (#279)
* Blacklist the CRB repository when Optional repo is not enabled

Previously the CRB repository was blacklisted always. Now we will
blacklist this one only when the Optional repository is not enabled
as the CRB is sort of counterpart of the Optional repo in RHEL 8
system.

Additionally, the filtering of blacklisted repositories in
the peseventsscanner actor is applied after repositories are mapped
from aliases to real repoids.

* Enable CRB repository when used during the upgrade

We are using the CRB repository during the upgrade when the Optional
repository was enabled on the source system. As this one
is counterpart to the Optional repo, enable it as well on the target
system (during the FirstBoot).

TODO: the actor is working with this one repo now, but would be nice
      to make it more robust to be used for any repository task
      in future. See the TODO msg in code.