RHSM: Retry refresh and release write actions on failure
Previously it happened that Satellite subscribed machines were failing
when trying to perform a task too fast. This patch introduces a way to
rate limit these tasks and have them retry on failures after sleeping
for some time.

Retry count with this patch is set to 5
Sleep time with this patch is set to 5 seconds

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>