Add actor CheckKdeGnome
A simple script detecting installed KDE desktop environment and its apps,
written in leapp framework. When started, the script will check whether
KDE is installed together with GNOME desktop to inform whether we can
inhibit the upgrade process. When both are installed, we need to inform
the user that KDE will be removed and GNOME will be used instead. If only
KDE is installed, we want to inhibit the upgrade process otherwise the
user will end up without a desktop. Also if both are installed,
but KDE is not used as default desktop, we further check most common
KDE apps to inform the user that those will be removed in case the user is using them.

Original script in bash written by jgrulich [0].
[0] https://gitlab.cee.redhat.com/jgrulich/rhel-upgrade