Fix traceback in dnfupgradetransaction if RHSM is skipped
Previously some change introduced removing certificates as specified by
the SourceRHSMInfo message.
However this message might not be available if RHSM is skipped.
In that case a None object will be returned by the next() call.
While trying to access the existing_product_certificates member of the
message, which actually is a None object, an AttributeError exception
happens.

This patch fixes this issue.

Fixes Rabbit: OAMG-515
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>