xfs: Provide information per mountpoint if XFS is present or not
Previously we did not report if XFS is present on a certain partition.
This was sufficient because did at most need only one overlayfs view.
However since we do want to be able to mount multiple partitions with
overlayfs and bindmount them together, we try to optimize the amount of
space needed for the scratch files, by not creating disk images for
partitions that do not require it.
This patch introduces the update for the information that is passed on.
A followup patch will take care of the actual mounting for the
additional partitions.

Addititonally this patch moves the actor from the checks phase to the
fact collection phase. It does not produce any checks or inhibitors. The
move is necessary as the information is already required in the facts
collection phase.

Fixes: https://github.com/oamg/leapp-repository/issues/334
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>