CheckKernelDrivers: List of blacklisted drivers mirrors the official RH docs (#336)
CheckKernelDrivers: List of blacklisted drivers mirrors the official RH docs

Previously the list of removed drivers contained all drivers we discover we cannot find on RHEL 8. The official list of removed drivers is much smaller. Probably those drivers we cannot find and are not included in the official list, are renamed or included in the kernel itself. So far, we start to blacklist just the drivers from the official list.

The others will be moved to separate file just for info. In case anyone find an issue related to one of thosedrivers, we can reflect that later.