overlayfs: Multi partition support
Previously leapp did not support more than /boot, /home and the / (root)
partitions. The problem lies in the way how overlay mounts work. This
patch introduces changes to the code base to allow for any partition
that is configured in /etc/fstab. Other parititions are not being
considered and currently expected to be irrelevant to the system's
upgrade.

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>