Fix a broken assertion in a test
The assertion was incorrectly rewritten in commit 548a3b83909c4. The
assertion is not looking for the expected data where it should. To
test the actual data, we need to reach far deeper into the object hierarchy.

As it is now, the assertion effectively checks that the list [False]
is True (which it is, because a non-zero length list is True). So the
assertion passes even though we don't want it to.