Generate correctly initrd on s390x arch
On the s390x arch is used DASD. To be able to create initrd correctly
we need to copy the /etc/dasd.conf into the RHEL 8 userspace and
add install it inside the dracut image explicitly as well.

The explicit installation is required because it is installed
automatically only in the case the `--hostonly` option is specified
when call dracut. But use of the option has other consequences,
e.g. many additional files we need now are removed. It's probably
because we generate the initrd file inside very limited bootstrap
of RHEL 8 adn dracut cannot detect the real setup of the host system.

Additionaly add quotes for LEAPP_ADD_DRACUT_MODULES envar to be
able to add multiple modules.

The list of files that should be installed in dracut can be modified
through LEAPP_DRACUT_INSTALL envar (not from cmdline, the var is set
inside the actor). By default, install just `systemd-nspawn`
(as before).

The solution implemented in the library is not perfect
and it is expected more robust implementation in future, that will
include consume of messages to be able to specify what files should
be installed.

Fix in generate-initram.sh:
  In the generate-initram.sh script have been several conditions
  incorrect as these checked only string instead of variables.