dracut: Ensure /usr is mounted in the initram container
Previously /usr wasn't mounted when we started leapp in the initram
disk. Since mount lives in /usr/bin/mount the call to mount -a failed
too and we ended up with a emergency console during the initram phase.

This patch introduces changes that /usr is mounted into the $NEWROOT
before we start leapp. To do so, we added a new script derived from the
98usrmount dracut module that mounts /usr into the $NEWROOT before we
start leapp.

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>