Do not filter events by installed packages
Chaining, man.

In a nutshell, an event for release A can introduce package P to be installed. If these events would be filtered beforehand, an event for a later release B that operates on package P wouldn't even make it to process_events.
Maybe we should refactor more of this actor, since it's pretty spaghetti-ish at places. Too bad the release is looming.

Also corrected tests.