Add check for installed multiple kernels on s390x
We discovered recently that removal of old kernels is not handled
correctly during the IPU. In case the maximum number of kernels
are installed, the oldest one is automatically uninstalled during
the rpm upgrade transaction.

This could lead into unbootable state on some machines, and the
bootloader configuration can be corrupted after the removal as well.
This is problem especially on s390x machines.

To prevent any related troubles during the IPU, inhibit the IPU
on s390x unless just one kernel is installed, until the issue will
be fixed correctly.