Removed UdevAdmInfo actor (#330)
Fix check of active kernel drivers: remove `udevadm info -e` feature

The feature creates false positives because it detects some drivers under different names, like
 "driver_a" as "driver-a".
See example:
 "virtio_pci" -> "virtio-pci"

Practical implication is, that even when we blacklist some drivers to inhibit upgrade when these are used, they will not be detected. Instead of that, we remove the feature. Negative of this step is, that upgrade will be inhibited when such driver will be loaded even in case it is not used. But rather some bothering of users by requested manual action than let them break the system.