s390x: add actor to inhibit when rd.znet in cmdline
Some s390x machines have broken networking after the upgrade
on RHEL 8. Currently we are not able to tell why and which machines
have this problem, but one of possible workrounds could be remove
of the rd.znet parameter from the kernel cmdline.

Problem is that such change can break networking as well, e.g.
in case a VLAN is set. Inhibit upgrade when rd.znet is detected
on kernel cmdline.

In case any VLAN is not detected, add hint to remove the rd.znet
parameter from the cmdline.

The actor is without tests because of the lack of time and it is
temporary solution until we fix it.