Rewrite PCIDeviceScanner and CheckKernelDrivers
The previous version of CheckKernelDrivers compared the list of drivers
removed in RHEL 8 against present kernel modules. The problem is,
drivers and modules don't always map 1:1 to each other and in some
cases, their names differ.

This commit reworks CheckKernelDrivers to work with driver names. Those
are now produced in PCIDeviceScanner, which is also reworked along with
the PCIDevices model, to which more fields are added.
WhitelistKernelModules and its model WhitelistedKernelModules are
removed, since they are no longer needed.