Inhibit upgrade when upgrade is unsupported
The upgrade is deemed unsupported if any LEAPP_DEVEL_* variables are
used or any experimental actors are enabled. When this happens, the
upgrade is inhibited.

This can be bypassed by setting LEAPP_UNSUPPORTED, at user's own risk.
A warning will be displayed warning when this happens (this is covered
by the framework).