Revert "Add "requests" package also to requirements (#410)"
This reverts commit bb7ac2c6d738ee0574171d8b47a3d3db73e49ef4.
That commit was an incorrect workaround on the leapp-repository side. It
should have been fixed on the leapp framework side.

The 'requests' package was missing when running leapp-repository unit tests. It was caused when leapp framework was being installed as a dependency by leapp-repository makefile, but without this 'requests' dependency of leapp framework itself. It's fixed by https://github.com/oamg/leapp/pull/587.