Makefile: specify chroot by COPR_CHROOT, optional
In some cases we need to build rpms just for specific chroot. E.g.
to make building faster as our rpms are noarch (so build just for
one arch...). Or in case of public COPR, build leapp-repository just
for epel, as there is no repository for Fedora yet.

Possible use:
  COPR_CHROOT=epel-7-x86_64 make copr_build