[RFE: beta|htb] repomap scanner: reflect product type
The new repomap file contains 2 additional fields:
  - src product type
  - dst product type
which can contain one these values:
  - ga
  - beta
  - htb

By default only 'ga' repositories should be used. But for devel
purposes it should be possible to use beta and htb repositories
as well. To change the product type, set envars:
  - LEAPP_DEVEL_SOURCE_PRODUCT_TYPE
  - LEAPP_DEVEL_TARGET_PRODUCT_TYPE
Invalid values are interpretted as 'ga'. Only repositories of
expected product type are used. Others are skipped.

The original model is unchanged. The new fields are used just for
filtering and there is no need to use them in any next actor.

The actor has been rewritten to be able to write more robust tests.
Unit-tests are completely rewritten, using generated data, instead
of crafted files (which have been dropped). The actor is not
producing the Report inhibitor message anymore as in such case,
the message was not showed because of crash of following actors.
Raise StopActorExecutionError with proper message instead.