[RFE: beta|htb] handle certificates
Improved the original map to choose the correct repositories
to reflect current situation:
  - removed variant: there is not any variant on RHEL 8
  - rename channel to 'product type' and instead of 'base', ...
    channels use products types: ga,beta,htb

Expecting that Beta and HTB repositories are irrelevant for already
GA released systems, so added just certs for RHEL 8.2 now.

Note: It is question whether we will need to handle in future EUS
as well, as another type of product. But probably we will be able
to access EUS repositories with the GA certificates if provided
by subscription. (someone should check that)

TODO: update tests