Optimize filter_out_pkgs_in_blacklisted_repos
Previously get_repositories_blacklisted has been called for every
package that has been marked to be installed. That call is very
expensive as it creates a temporary set with all blacklisted
repositories even though the data could be only generated once and
reused. This patch ensures it is only done once.

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>