Use real path for systemd-nspawn in initramfs
An alternative to #439, this is an attempt to actually solve the bug
rather than inhibiting.
The core issue is that if /bin is a link to an absolute path, the
resolution of /bin/systemd-nspawn to /usr/bin/systemd-nspawn fails. So
why not just directly use the latter?

(Pretty naive. Several people told me this shouldn't break anything,
but I don't trust that.)