Modify IPUConfig: add kernel field
Add info about the originally booted kernel. IOW, after boot into
the upgrade initrd or upgraded system, the information is persisent
and contains still the kernel version and release detected on the
source system when the workflow started. E.g.:
    4.14.0-100.el7a.aarch64

This is prerequisite to be able to check whether system is RHEL
or RHEL-ALT.