Add an actor for modifying the multipath configuration files
This actor changes the the multipath configuration files so that they
will work in RHEL8, with as close a configuration to the RHEL7 version
as possible.

The actor silently comments out all removed configuration options,
specifically "hw_str_match", "ignore_new_boot_devs",
"new_bindings_in_boot" and "unpriv_sgio", since these all are no longer
necessary in RHEL8. It changes the option "detect_path_checker" to
"detect_checker", since the name has changed, and it removes
"hardware_handler" and "pg_timeout" from the overrides section, since
these no longer exist and never did anything in RHEL7.

If a features option contains "queue_if_no_path" it is removed, since it
has been deprecated in the config, and adds a "no_path_retry queue" option
instead, if the "no_path_retry" option did not already exist. This has
not change on the effective configuration.

If a device section with "all_devs" exist, the configuration values are
added to an overrides sections, since this replaces "all_devs" device
configs in RHEL8. If an overrides section already exists in the RHEL7
config, its values have priority over the all_devs values.

The actor also changes the values of "retain_attached_hw_handler",
"detect_checker", "detect_prio", and "reassign_maps" to the RHEL8
defaults, because this is necessary for proper functioning, but warns
the user of this.

Finally, if the path checker is set in the defaults section to something
other than "tur", it is not obvious how to upgrade the configuration,
and so the actor inhibits it.

v2 & v3: fix some lint issues
v4: fix missing config handling in multipathutil.py and add a test for
    it.
v5: Fix _nothing_to_do() from ignoring configs that need work, and add
    tests to check for files that only need one thing changed.
v6: Cleanup _nothing_to_do(), cleanup some checks, and change double
    quoted strings to single quoted strings