verifycheckresults: Do not report inhibitors as errors
Previously there was no way to distinguish a normal error from an
inhibitor related error. Having the information will help in improving
the user experience by giving the user a hint where to look for more
information in regards to the error (preupgrade report).

verify check results now will just request the stop of the execution
instead of reporting errors.

For this patch leapp-framework version 1.1 is required.

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>