[RFE custom repos 1/9] add actor to scan the default custom repo file
The actor scan the custom repo file that user could use to define
own repositories that should be used during the IPU as well.

The expected path of the file is:
    /etc/leapp/files/leapp_upgrade_repositories.repo

These repositories will be used automatically for the IPU despite
the enable/disable settings. It's required the file and defined
repositories has to be valid.

If the file doesn't exist, nothing happens.