[RFE custom repos 2/9] Add the CustomTargetRepositoryFile model
The model is used to share information about custom repo files that
should be extra handled during the IPU. Basically we have just one
file that we want to use for these purposes now, but I haven't found
good library for the constant and I haven't wanted to duplicate the
string several times. As well, maybe someone will want to handle own
repofile. For that reasons, I rather created the new model.

Note: Feel free to drop it and hardcode the path.