Provide upgrade inhibitor for ipa-server
IdM (ipa-server) does not support in-place upgrade from RHEL 7 to 8.
Instead IdM servers must be upgraded by creating a new RHEL 8 machine
and installing a replica.

The new IpaScanner actor scans for the system for IdM packages and
client/server instances. The new CheckIPAServer prohibits upgrades when
an IdM server is installed and warns the user when ipa-server package is
present.

Signed-off-by: Christian Heimes <cheimes@redhat.com>