[RFE custom repos 7/9] yum/dnf: disable subscription-manager plugin if rhsm skpped
This is part of the scenario, when user wants to use only custom
yum/dnf repositories (rhsm is skipped), but the system is subscribed
in the same time.

YUm by default uses the subscription-manager plugin; DNF uses it when
configured. If the plugin is loaded, the redhat.repo file will be
regenerated (e.g. it could lead into unexpected duplicated). As well,
maybe it could happen that an error will be raised because of troubles
to obtain expected data (e.g.). So we disable it).