Implement generic CurrentActorMocked
Despite being one of the most mocked objects in our tests, current_actor
didn't have a united mock implementation so far.
This commit implements a generic, fully configurable mocked actor for
use in all repository unit tests. Possible extension points that are in
the real actor class definition are marked as not implemented for
clarity.