make actor tests name unique
This is needed because pytest is not able to run all tests in couple
of files with the same name. To ensure we are able to run all tests
with pytest, let's rename those files. See the good practice:
    https://docs.pytest.org/en/latest/goodpractices.html

The new practice for the naming of tests is like this:
   - unit_test_<actordir>.py
   - component_test_<actordir>.py

Or instead of actor dir, use the actor name. The goal is to ensure,
you will not have to spend minutes by searching the name that
is still free.