Rename actor private libs and update the code due to pytest
This is analogous to the previous commit. Pytest is not able to
run all tests properly with duplicate filenames of libraries.
Similar recommendation is used here for naming; instead
of `library.py`, call your library like `<actordir>.py` whatever
name that will contain uniq actor string.

In addition, in this PR:
- many unused imports are dropped
- some refactoring happens as well to make linter happy and improve
  readability
- Make paths of tests resources not dependent of cwd
  - done in order do not modify cwd when running actors tests since
    pytest rely on constant cwd equal repo root
  - unify test data pathes so it drives from the module path
  - see e.g. test_prepare_yum_config