mounting library: secure the isolated actions
Previously it was possible by mistake do actions out of the isolated
environment, e.g.
    context.copy_to('/path/src', '../../path/dst')

In such a case, the dst path is out of the basedir of the isolated
environment. To prevent this, update the IsolatedActions.full_path
method.

Currently, the solution transfer the path "inside the env"
to absolute path and then the join of paths is applied. As it is
expected to use always absolute paths, it should not be problem
in general.

Warning:
  But in case someone uses a relative path, it generates
  completely different path than expected. If needed, it
  can be replaced by solution where relative paths are unchanged,
  but in such a case, the method has to raise an exception when
  the path is in real out of the environment.