Add localrepoinhibit actor
The purpose is to inhibit the upgrade if the local repository
was found (which has baseurl starting with file:///).

The new model TMPTargetrepossitoriesFacts was introduced, due to the limitations
in UsedTargetRepositories and TargetRepositories models.

UsedTargetRepositories has only repoids data. TargetRepositories has baseurls of repos,
but not in all cases (i.e. urls are missing if the repo introduced inside the yum.repo.d dir).