[POC] initram networking
Adding initial basic networking support for the initram phase.
Controlled by the LEAPP_INITRAM_NETWORK environment variable which must
be set to either `scripts` or `network-manager` to choose between the
legacy or NetworkManager based dracut modules.

Recommended to use is the network-manager part at this moment as it will
take care of most of things including default routing etc.

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>