detectgrubconfigerror: Fix false positive on comments, add tests
The false positive occured when GRUB_CMDLINE_LINUX was followed by a
comment. In a later stage, it caused the AddUpgradeBootEntry actor to
break the GRUB config and the booting process.

Added tests include the /etc/default/grub config file on which this
behavior occured, according to the bug report:

https://bugzilla.redhat.com/show_bug.cgi?id=1872356#c0