Remove invalid XFS mount options from target system fstab
Several XFS mount options have been removed from the kernel between
RHEL 7 and 8. On RHEL 7, the options are deprecated and have no effect.
However, on RHEL 8, their occurence will lead to an unbootable machine.
If this actor detects any of the options in /etc/fstab, it removes them
during the upgrade. The user is notified about this in the preupgrade
report and a "read-only" copy of the new /etc/fstab is provided for
inspection.

The list of removed options is taken from xfs(5) on RHEL 8.

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=1870813