Inhibit upgrade if multiple kernel-debug pkgs are installed
When multiple kernel-debug packages are present, leapp can't process
the transaction in some cases, so we inhibit the upgrade and present
the user with hints on removing the surplus debug kernels.

Note that the added actor is basically a search-and-replace copy of
CheckInstalledDevelKernels, which solves the same problem for a
different package. I don't see many reasons to merge them into one.

Derived from: commit 2ce678fe8613c0d96ff720b7374466695aa6dc42 (#457)