checkinstalledkernels: detect kernel-rt and improve comparisons (deps changed)
RHEL Real Time (RT) systems boot using kernel-rt rpm. But still,
even a kernel rpm can be installed. So now, detect whether we are
booted via kernel-rt or not and do a comparison only in regarding
the expected package name.

Unfortunately, this make comparisons between releases more difficult
as we need to check the whole release, including alphanumeric
characters in release, e.g.:
        kernel-rt-3.10.0-1160.2.1.rt56.1133.el7.x86_64

Previously we worked just with numbers. To make the check valid and
more robust, let's use rpm.labelCompare(). We are sure this library
is on every RHEL 7 system as dnf requires it (and probably yum too).
Let's use it as it resolves all our current troubles right now and
makes whole comparison more reliable.

In this implementation I am skipping comparison with epoch as
there is no (and will not be) kernel with bumped epoch.

Important:
   Regarding the use of the rpm module, someone could expect added
   dependency on rpm in python or spefile. Currently, nothing like
   that. In case case of python (pip) dependencies, there is not
   python module presented via pip as the module is written
   in CPython. Only way to install it is via packages. However,
   on RHEL 7 you can find just rpm module for Python2 and newer
   systems contain just rpm module for Python3. So there is no way
   to prepare virtual pyenv with specific rpm modules and the system
   one has to be used.

   Changes in the specfile are not needed as the related rpm is
   always installed and cannot be removed from systems based on RHEL
   and Fedora.