Fixing log storing after upgrade in initramdisk
Previously, when running leapp upgrade during the initram disk, log
files weren't saved if /var/log was residing on a different partition.
This PR fixes this issue by running the saving process of the log file
in a container and mounting all partition before hand. The same approach
is used for running leapp, options to systemd-nspawn are shared through
a global environment variable.

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>