libraries.common.config.version: add detection of RHEL RT
The is_rhel_realtime function returns True only if the original
system was booted using preemptive kernel (kernel-rt) and when the
system is RHEL.

The function is currently naive a little bit. The realtime kernel
detection is based on the '.rt' substring in the release of kernel
- which we know it will be there always only in case of the kernel-rt
rpm. However, in future the test should be based rather on properties
 of kernel (preemptive-rt).