Influence creation of the targetuserspace container
This is part of wider changes around the targetuserspace container
and initramfs for upgrade and target system.

We discovered already several cases when the targetuserspace
container needs to be dynamically modified based on the host system.
This commit provides possibility to influence the creation of the
targetuserspace container via TargetUserspacePreupgradeTasks messages.
This commit covers the following use-cases:
  - copy a file A (from host) to B (to the targetuserspace)
  - install rpms inside the container

These use-cases should cover current needs to:
  - be able to connect to a server with repositories
    (e.g. the /etc/host file could be needed)
  - be able to calculate the upgrade transaction
    (e.g. the dnf, rpm, ... packages of the target system need to be
    installed)

Related actors have been updated to use the new functionality too.

Changes in Models
=================
- added new models:
  - TargetUserspacePreupgradeTasks
        to influence content of the target userspace container
        when it is created
- deprecated models:
  - RequiredTargetUserspacePackages
        replaced by TargetUserspacePreupgradeTasks