Influence the generation of the upgrade initramfs
This is part of wider changes around the targetuserspace container
and initramfs for upgrade and target system.

We have already bunch of hardcoded stuff around the work with the
targetuserspace container just to provide content required
to generate the working upgrade initramfs. This all was added when
we discovered we need it to be able to generate the initramfs or
to add additional stuff into the initramfs to be able to boot.
The same situation occurs for end-users who need to (e.g.) add
own kernel drivers into the upgrade (and possibly target) initramfs
and they couldn't do that because of the missing mechanism to
dynamically affect the generation of the ugprade initramfs.

This commit provides mechanisms to:
  - affect the targetusperspace container before the generation of
    the upgrade initramfs
    - copy a file A (from host) to B (to the targetuserspace)
    - installation of required rpms inside the targetuserspace
  - affect generation of the upgrade initramfs
    - install specific files into the initramfs
    - use a specified dracut module
      - the dracut module can be stored on different path. In such
        case, the dracut module will be automatically copied into the
        targetuserspace under the expected expected directory to be
        discovered by dracut

The original hardcoded fixes around Multipath and DASD are kept
in the actor now. They will be removed in the following commits
during update of related actors to use the new functionality.

Changes in Models
=================
- added new models:
  - UpgradeInitramfsTasks
        influence content in the upgrade initramfs
  - TargetUserspaceUpgradeTasks
        to influce content of the target userspace container before
        the upgrade initramfs is created
- deprecated models:
  - RequiredUpgradeInitramPackages
        replaced by TargetUserspaceUpgradeTasks
  - UpgradeDracutModule
        replaced by UpgradeInitramfsTasks