Multipath: create task to copy config files into the targetuserspace
Previously the coping of multipath configuration files was hardcoded
inside the actor responsible for generation of the upgrade initramfs.
As the new functionality to influence content of the upgrade initramfs
and the targetuserspace is provided, split this functionality out
into multipath related actors.

The current solution is not ideal (see the TODO note) regarding other
stuff that was discovered around the multipath. The current
functionality is expected to be updated in future. Now, it behaves
in the same way from the user POV as before.