Influence the generation of the target initramfs
This is analogy to the previous implementation of the mechanism to
influence generation of the upgrade initramfs (but work with the
targetuserspace, which is not relevant anymore). Use-cases
and reasons for the influence of the upgrade initramfs are pretty
same, just we need it this time for the target initramfs.

This commit provides mechanisms to:
  - affect generation of the target initramfs
    - install specific files into the initramfs
    - use a specified dracut module
      - the dracut module can be stored on different path. In such
        case, the dracut module will be automatically copied into
        the expected expected directory to be discovered by dracut

The actors (persistentnetnamesconfig) is updated as well regarding
trivial change in the actor.

Changes in Models
=================
- added new models:
  - TargetInitramfsTasks
        influence content in the target initramfs
- deprecated models:
  - InitrdIncludes
        replaced by TargetInitramfsTasks

TODO:
  Missing correct testing with the coping of dracut modules.
  And the coping of dracut modules needs to be improved as well.