Move and rename initramfs-related actors
The initrdinclude and initramdiskgenerator actors are not intuitive
names and it's hard to guess what is doing which of these actors.
This was mainly because of original nature of the actors. However,
with the new extended functionality, it is better to rename these
actors.

Move the related actors into the initramfs directory. The rename:
    - initramdiskgenerator -> upgradeinitramfsgenerator
    - initrdinclude        -> targetinitramfsgenerator
Actors's libraries and test files have been renamed as well
to respect new names, following recommendation in official guidelines.