persistentnetneamesconfig: crash on non-existent iface
We have to deal with situation when:
  - we detect a network device on src system
  - we cannot detect the network device inside the upgrade env

This happens very easily, when the nerwork card need specific driver
which is by default not included in the initramfs. As we are not
able to deal with any possible driver (especially from the other
vendors), let's rather ensure the actor doesn't crash in such a case
and rather prints a warning message(s).

NOTE:
We are not able to load the IB card properly in the upgrade initramfs
as so far, it needs working networking otherwise interfaces on the IB
are not loaded by kernel (even when IB is detected - long story..).